add default prompt for styles, move invetory to resources and overview report to tool
