--- brewfather_mcp/inventory.py
@@ -1,1 +1,0 @@
-
