+++ src/brewfather_mcp/api.py
@@ -1,5 +1,4 @@
 import os
-import typing
 
 import httpx
